Cleanup routes - server-side
* Gets rid of v2 prefix
* Moves files around for cleaner structure
* Misc.
