+++ server/package.json
@@ -43,9 +43,9 @@
     "express-fileupload": "^1.2.1",
     "fast-json-stringify": "^5.9.1",
     "fast-xml-parser": "^4.3.5",
-    "fastify": "^4.24.3",
+    "fastify": "^4.26.0",
     "fastify-plugin": "^4.5.1",
-    "fastify-print-routes": "^2.2.0",
+    "fastify-print-routes": "^3.1.1",
     "fastify-type-provider-zod": "^1.1.9",
     "fluent-ffmpeg": "^2.1.2",
     "lodash-es": "^4.17.21",
