--- server/src/api/v2/plexServersApiV2.ts
@@ -1,62 +1,0 @@
-import { BaseErrorSchema } from '@tunarr/types/api';
-import { isError, isNil } from 'lodash-es';
-import z from 'zod';
-import { PlexServerSettings } from '../../dao/entities/PlexServerSettings.js';
-import createLogger from '../../logger.js';
-import { PlexApiFactory } from '../../plex.js';
-import { RouterPluginAsyncCallback } from '../../types/serverType.js';
-import { wait } from '../../util.js';
-
-const logger = createLogger(import.meta);
-
-// eslint-disable-next-line @typescript-eslint/require-await
-export const plexServerApiV2: RouterPluginAsyncCallback = async (fastify) => {
-  fastify.addHook('onError', (req, _, error, done) => {
-    logger.error(req.routeConfig.url, error);
-    done();
-  });
-
-  fastify.get(
-    '/plex/status',
-    {
-      schema: {
-        querystring: z.object({
-          serverName: z.string(),
-        }),
-        response: {
-          200: z.object({
-            healthy: z.boolean(),
-          }),
-          404: BaseErrorSchema,
-          500: BaseErrorSchema,
-        },
-      },
-    },
-    async (req, res) => {
-      try {
-        const server = await req.entityManager
-          .repo(PlexServerSettings)
-          .findOne({ name: req.query.serverName });
-
-        if (isNil(server)) {
-          return res.status(404).send({ message: 'Plex server not found.' });
-        }
-
-        const plex = PlexApiFactory.get(server);
-
-        const s = await Promise.race([
-          plex.checkServerStatus().then((res) => res === 1),
-          wait(15000).then(() => false),
-        ]);
-
-        return res.send({
-          healthy: s,
-        });
-      } catch (err) {
-        return res.status(500).send({
-          message: isError(err) ? err.message : 'Unknown error occurred',
-        });
-      }
-    },
-  );
-};
